backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.9",
    "psycopg2>=2.9.11",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0",
    "argon2-cffi>=25.1.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
]
